requirements.txt
django>=4.2
djangorestframework>=3.14
mongoengine>=0.27
django-cors-headers>=4.3
groq>=0.9
python-dotenv>=1.0
bcrypt>=4.1
PyJWT>=2.8
channels>=4.0
daphne>=4.0

firebase-admin>=6.0